requirements.txt
psycopg2
botcity-email-plugin